backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
DateTime==5.5
fastapi==0.122.0
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6
numpy==2.3.5
orjson==3.11.4
packaging==24.2
pandas==2.3.3
pillow==12.0.0
pinecone==8.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
zope.interface==8.1.1
